+++ packages/koot/core/webpack/defaults/isomorphic.server.dev.js
@@ -25,7 +25,8 @@ const factoryConfig = async ({
             filename: '[name].js',
             chunkFilename: 'chunk.-_-_-_-_-_-[chunkhash]-_-_-_-_-_-.js',
             path: `${pathRun}/${common.outputPath}/server`,
-            publicPath: `${publicPath}/`
+            publicPath: `${publicPath}/`,
+            pathinfo: false
         },
         plugins: [
             new webpack.DefinePlugin({